Add UriBuilder.fromUri(), toString()
