+++ test/uri_test.dart
@@ -24,6 +24,12 @@ main() {
           Uri.parse('ftp://userInfo@host:42/path?q=v#fragment'));
     });
 
+    test('should build a Uri from a Uri', () {
+      var uri = Uri.parse('http://example.com:8080/path?a=b#fragment');
+      var builder = new UriBuilder.fromUri(uri);
+      expect(builder.build().toString(), uri.toString());
+    });
+
   });
 
   group('UriMatch', () {
